PHP master project files

+++ mysql/functions.php
@@ -1,20 +1,23 @@
 <?php include "db.php"; ?>
 <?php
 
+
 function createRows()
 {
     if (isset($_POST['submit'])) {
         global $connection;
+
         $username = $_POST['username'];
         $password = $_POST['password'];
 
-        /*$connection = mysqli_connect('localhost', 'root', '', 'loginapp' );
+        $username = mysqli_real_escape_string($connection, $username);
+        $password = mysqli_real_escape_string($connection, $password);
 
-        if($connection) {
-            echo "We are connected";
-        } else {
-            die("Database connection failed");
-        }*/
+
+        $hashFormat = "$2y$10$";
+        $salt = "iusesomecrazystrings22";
+        $hashF_and_salt = $hashFormat . $salt;
+        $password = crypt($password, $hashF_and_salt);
 
         $query = "INSERT INTO users(username,password) ";
         $query .= "VALUES ('$username', '$password')";
@@ -22,12 +25,31 @@ function createRows()
         $result = mysqli_query($connection, $query);
         if (!$result) {
             die('Query FAILED' . mysqli_error());
-        }
-        else{
-           echo "Record Created";
+
+        } else {
+
+            echo "Record Create";
+
         }
     }
 }
+
+
+function readRows()
+{
+    global $connection;
+    $query = "SELECT * FROM users";
+    $result = mysqli_query($connection, $query);
+    if (!$result) {
+        die('Query FAILED' . mysqli_error());
+    }
+
+    while ($row = mysqli_fetch_assoc($result)) {
+
+        print_r($row);
+    }
+}
+
 
 function showAllData()
 {
@@ -40,44 +62,58 @@ function showAllData()
 
     while ($row = mysqli_fetch_assoc($result)) {
         $id = $row['id'];
+
         echo "<option value='$id'>$id</option>";
     }
 }
 
+
 function UpdateTable()
 {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if (isset($_POST['submit'])) {
 
-    $query = "UPDATE users SET ";
-    $query .= "username = '$username', ";
-    $query .= "password = '$password' ";
-    $query .= "WHERE id = $id ";
-    $result = mysqli_query($connection, $query);
-    if (!$result) {
-        die("QUERY FAILED" . mysqli_error($connection));
-    }
-    else {
-        echo "The record was successfully updated";
+        global $connection;
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $id = $_POST['id'];
+
+        $query = "UPDATE users SET ";
+        $query .= "username = '$username', ";
+        $query .= "password = '$password' ";
+        $query .= "WHERE id = $id ";
+
+        $result = mysqli_query($connection, $query);
+        if (!$result) {
+
+            die("QUERY FAILED" . mysqli_error($connection));
+        } else {
+
+            echo "Record Updated";
+
+        }
     }
 }
 
+
 function deleteRows()
 {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if (isset($_POST['submit'])) {
+        global $connection;
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $id = $_POST['id'];
 
-    $query = "DELETE FROM users ";
-    $query .= "WHERE id = $id ";
-    $result = mysqli_query($connection, $query);
-    if (!$result) {
-        die("QUERY FAILED" . mysqli_error($connection));
-    }
-    else{
-        echo "The record was successfully deleted";
+        $query = "DELETE FROM users ";
+        $query .= "WHERE id = $id ";
+
+        $result = mysqli_query($connection, $query);
+        if (!$result) {
+
+            die("QUERY FAILED" . mysqli_error($connection));
+        } else {
+
+            echo "Record Deleted";
+
+        }
     }
 }
